deVirgo_1.py
'''
deVirgo 是一个基于零知识证明的隐私保护协议，它可以用于实现匿名投票、隐私搜索等应用。下面是一个使用 Python 实现的 deVirgo 零知识证明的示例代码。

在这个示例中，我们假设有一个 Bob 和一个 Alice，Bob 想要向 Alice 证明自己知道一个字符串 $s$，但是不想让 Alice 知道具体的字符串是什么。

'''

import random
import hashlib
from Crypto.Util import number
from typing import List, Tuple

# 定义一些常量
N_BITS = 1024  # 随机数的位数
G = 2  # 循环群的生成元
P = int(
    'FFFFFFFFFFFFFFFFC90FDAA22168C234C4C6628B80DC1CD129024E088A67CC74020BBEA63B139B22514A08798E3404DDEF9519B3CD3A431B302B0A6DF25F14374FE1356D6D51C245E485B576625E7EC6F44C42E9A63A3620FFFFFFFFFFFFFFFF',
    16)


# 定义辅助函数
def hash_string(string: str) -> int:
    """
    计算一个字符串的哈希值。
    """
    hash_object = hashlib.sha256(string.encode())
    return int.from_bytes(hash_object.digest(), 'big')


def gen_rand_num() -> int:
    """
    生成一个随机的大整数。
    """
    return number.getPrime(N_BITS)


class deVirgo:
    def __init__(self):
        # 生成一个随机数 p，p 是一个大质数
        self.p = gen_rand_num()

        # 计算循环群的阶
        self.q = (self.p - 1) // 2

        # 生成一个随机数 g，g 是循环群的生成元
        self.g = pow(G, (self.p - 1) // self.q, self.p)

        # 生成一个随机数 x，作为 Bob 的私钥
        self.x = gen_rand_num()

        # 计算 Bob 的公钥 y
        self.y = pow(self.g, self.x, self.p)

    def gen_proof(self, s: str) -> Tuple[int, int]:
        """
        生成 deVirgo 零知识证明。
        """
        # 将字符串 s 转换成一个整数
        m = hash_string(s)

        # 生成一个随机数 r，作为零知识证明的随机挑战
        r = gen_rand_num()

        # 计算挑战 c
        c = pow(self.g, r, self.p) % self.q

        # 计算响应 z
        z = (r - self.x * c) % self.q

        # 返回挑战和响
        return (c, z)

    def verify_proof(self, s: str, proof: Tuple[int, int]) -> bool:
        """
        验证 deVirgo 零知识证明。
        """
        # 将字符串 s 转换成一个整数
        m = hash_string(s)

        # 解析出挑战和响应
        c, z = proof

        # 计算 u1 和 u2
        u1 = (pow(self.g, z, self.p) * pow(self.y, c, self.p)) % self.p
        u2 = pow(self.g, m, self.p)

        # 计算 v
        v = pow(u1, hash_string(str(u2)), self.p)

        # 验证零知识证明是否合法
        return c == v % self.q


'''
这个示例中，我们创建了一个 deVirgo 实例，并使用它来生成一个字符串 s 的零知识证明。

然后我们验证这个零知识证明是否合法。
如果证明是合法的，我们将会输出 "The proof is valid!"。
否则，我们将会输出 "The proof is invalid."。
'''

# 创建一个 deVirgo 实例
dv = deVirgo()

# 生成一个字符串 s，我们将证明我们知道 s
s = "Hello World!"

# 生成零知识证明
proof = dv.gen_proof(s)

# 验证零知识证明
result = dv.verify_proof(s, proof)

if result:
    print("The proof is valid!")
else:
    print("The proof is invalid.")
